requirements.txt
pinecone-client
firstbatch